issue(Publisher QoS): removing publisher setPriority and setCongestionControl during runtime.
Done after changes on Zenoh from https://github.com/eclipse-zenoh/zenoh/pull/1124
