+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/jni/JNIPublisher.kt
@@ -14,9 +14,6 @@
 
 package io.zenoh.jni
 
-import io.zenoh.*
-import io.zenoh.prelude.CongestionControl
-import io.zenoh.prelude.Priority
 import io.zenoh.value.Value
 
 /**
@@ -53,51 +50,6 @@ internal class JNIPublisher(private val ptr: Long) {
     fun close() {
         freePtrViaJNI(ptr)
     }
-
-    /**
-     * Set the congestion control policy of the publisher.
-     *
-     * This function is not thread safe.
-     *
-     * @param congestionControl: The [CongestionControl] policy.
-     * @return A [Result] with the status of the operation.
-     */
-    fun setCongestionControl(congestionControl: CongestionControl): Result<Unit> = runCatching {
-        setCongestionControlViaJNI(congestionControl.value, ptr)
-    }
-
-    /**
-     * Set the priority policy of the publisher.
-     *
-     * This function is not thread safe.
-     *
-     * @param priority: The [Priority] policy.
-     * @return A [Result] with the status of the operation.
-     */
-    fun setPriority(priority: Priority): Result<Unit> = runCatching {
-        setPriorityViaJNI(priority.value, ptr)
-    }
-
-    /**
-     * Set the congestion control policy of the publisher through JNI.
-     *
-     * This function is NOT thread safe.
-     *
-     * @param congestionControl The congestion control policy.
-     * @param ptr Pointer to the publisher.
-     */
-    private external fun setCongestionControlViaJNI(congestionControl: Int, ptr: Long)
-
-    /**
-     * Set the priority policy of the publisher through JNI.
-     *
-     * This function is NOT thread safe.
-     *
-     * @param priority The priority policy.
-     * @param ptr Pointer to the publisher.
-     */
-    private external fun setPriorityViaJNI(priority: Int, ptr: Long)
-
 
     /** Puts through the native Publisher. */
     @Throws(Exception::class)
